railway-audio-checker/pyproject.toml
[project]
name = "sonar-audio-checker"
version = "0.1.0"
description = "Audio quality and copyright detection service for SONAR"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydub>=0.25.1",
    "librosa>=0.10.0",
    "numpy>=1.24.0",
    "pyacoustid>=1.3.0",
    "httpx>=0.25.0",
    "python-multipart>=0.0.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
